Move feature_compute files.
