+++ cpp/common/features/feature_compute_funcs.cpp
@@ -16,7 +16,7 @@
 
 #include <common/hash_utils.h>
 #include <common/arrow/arrow_helpers.h>
-#include <serving/feature_compute_funcs.h>
+#include <common/features/feature_compute_funcs.h>
 #include <common/utils.h>
 
 #include <fmt/format.h>
