+++ cpp/common/features/feature_compute_funcs.h
