+++ cpp/tests/serving/feature_compute_exec_test.cpp
@@ -15,7 +15,7 @@
 //
 
 #include <common/arrow/arrow_helpers.h>
-#include <serving/feature_compute_exec.h>
+#include <common/features/feature_compute_exec.h>
 #include <serving/test_utils.h>
 #include <common/types.h>
 #include <common/utils.h>
