+++ cpp/tests/serving/schema_parser_test.cpp
@@ -19,7 +19,7 @@
 #include <boost/asio/use_future.hpp>
 
 #include <common/arrow/arrow_helpers.h>
-#include <serving/feature_compute_exec.h>
+#include <common/features/feature_compute_exec.h>
 #include <serving/schema_parse.h>
 #include <serving/test_utils.h>
 #include <common/utils.h>
